refactor: rename smsdelivery.service to smsdelivery.services
